feat(#809): create shared @bluecollar/types package to eliminate type drift
- Add packages/types with canonical DTOs shared by api and app
- Update app types/index.ts to re-export from @bluecollar/types
- Wire @bluecollar/types as workspace:* dependency in api and app
- Add .github/workflows/type-check.yml to catch type drift in CI
