+++ package.json
@@ -10,6 +10,7 @@
     "dev": "concurrently \"pnpm --filter @bluecollar/api dev\" \"pnpm --filter @bluecollar/app dev\"",
     "build": "pnpm --filter @bluecollar/api build && pnpm --filter @bluecollar/app build",
     "test": "pnpm --filter @bluecollar/api test && pnpm --filter @bluecollar/app test",
+    "type-check": "pnpm --filter @bluecollar/types type-check && pnpm --filter @bluecollar/api tsc --noEmit && pnpm --filter @bluecollar/app type-check",
     "lint": "pnpm --filter @bluecollar/api lint && pnpm --filter @bluecollar/app lint",
     "format": "prettier --write \"packages/**/*.{ts,tsx,js,json,md}\"",
     "docker:up": "docker compose up -d",
