+++ packages/api/package.json
@@ -71,6 +71,7 @@
     "zod": "^3.23.0"
   },
   "devDependencies": {
+    "@bluecollar/types": "workspace:*",
     "@faker-js/faker": "^10.2.0",
     "@stryker-mutator/core": "^8.2.6",
     "@stryker-mutator/vitest-runner": "^8.2.6",
