+++ packages/types/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@bluecollar/types",
+  "version": "1.0.0",
+  "description": "Shared TypeScript types and DTOs for BlueCollar API and App",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "type-check": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.4.0"
+  },
+  "license": "MIT"
+}
